# docs/dev/api_spec_core_v1_v1.1.yaml
openapi: 3.0.3
info:
  title: ReMedi API — Phase 1 (Core)
  version: "0.4.0"
  description: >
    Phase 1(코어) 범위의 API 명세. Auth / 문서인식(공통모듈) / 복약 / Chat(RAG) / Push구독 / DUR 스크리닝을 포함한다.
    인증은 JWT(Access/Refresh Token, `Authorization: Bearer` 헤더)로 처리하며(2026-07, httpOnly 쿠키에서 변경),
    인증이 필요한 엔드포인트는 security: bearerAuth로 표기한다.
  x-changelog:
    - version: "0.1.0"
      date: "2026-07"
      changes: "초안 — Phase 1 Core 범위 (Auth/문서인식/복약/Chat/Push구독), 인증은 httpOnly 쿠키 기반"
    - version: "0.2.0"
      date: "2026-07-06"
      changes: "인증 방식을 httpOnly 쿠키에서 JWT(Authorization: Bearer)로 전환 — securitySchemes, AuthTokenResult, /auth/refresh, /auth/logout 갱신"
    - version: "0.3.0"
      date: "2026-07-07"
      changes: "복약 알림 일정(T-NTFY) CRUD 추가 — /notifications/schedules, NotificationSchedule* 스키마"
    - version: "0.4.0"
      date: "2026-07-14"
      changes: "DUR 3단계 스크리닝(T-MED-14) 추가 — /dur/screening/basic|interaction|ingredient, Dur* 스키마. 로그인 불필요(profile_id 미사용)."
servers:
  - url: https://api.remedi.app/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignupRequest:
      type: object
      required: [email, password, nickname, name, phone, birth_date, gender, email_verified_token, phone_verified_token]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        nickname: { type: string }
        name: { type: string }
        phone: { type: string, example: "010-1234-5678" }
        birth_date: { type: string, format: date }
        gender: { type: string, enum: [M, F] }
        email_verified_token: { type: string, description: "이메일 인증 완료 후 발급된 임시 토큰" }
        phone_verified_token: { type: string, description: "휴대폰 인증 완료 후 발급된 임시 토큰" }
    AuthTokenResult:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        message: { type: string, example: "로그인 성공" }
        access_token: { type: string, description: "JWT Access Token — 이후 요청의 Authorization: Bearer 헤더에 실어 보낸다" }
        refresh_token: { type: string, description: "JWT Refresh Token — /auth/refresh 호출 시 사용" }
        token_type: { type: string, example: "bearer" }
...
    ChatSessionCreateResult:
      type: object
      properties:
        session_id: { type: string, format: uuid }
    ChatMessageRequest:
      type: object
      required: [message]
      properties:
        message: { type: string }
        input_type: { type: string, enum: [text, voice], default: text }
...
    ChatHistoryItem:
      type: object
      properties:
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        sources:
          type: array
          nullable: true
          description: "T-LLM-7-3-2: 답변 생성에 쓰인 DUR/논문 출처 목록. user 메시지는 항상 null"
          items:
            type: object
            properties:
              name: { type: string }
              url: { type: string, nullable: true }
        disclaimer: { type: string, nullable: true, description: "의료 관련 답변에만 붙는 면책 문구. 없으면 null" }
        created_at: { type: string, format: date-time }

    # ---------- Push 구독 ----------
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: 이메일 회원가입 완료 (T-AUTH-1)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: 가입 완료, 응답 바디로 Access/Refresh 토큰 발급, 즉시 홈 진입 가능
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenResult' }
        '400': { description: 인증 미완료 상태로 가입 시도 }
        '409': { description: 닉네임 중복 }
...
  /auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃 (T-AUTH-4)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                push_subscription_endpoint: { type: string, description: "현재 기기 구독 해지용" }
      responses:
        '200': { description: "서버 처리 실패해도 클라이언트는 저장된 토큰 삭제로 로그아웃 완료 처리" }

  # ================= 문서 인식 (공통모듈) =================
...
  /chat/sessions:
    post:
      tags: [Chat]
      summary: 새 대화 세션 시작
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: 세션 생성됨 (Redis에 빈 스레드 생성)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatSessionCreateResult' }
  /chat/sessions/{session_id}/messages:
    get:
      tags: [Chat]
      summary: 현재 세션의 대화 이력 조회 (Redis 우선, 없으면 빈 배열)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: session_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 대화 이력
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChatHistoryItem' }

    post:
      tags: [Chat]
      summary: 메시지 전송 및 스트리밍 응답 수신 (T-LLM-2)
      description: >
        처리 순서: (1) SafetySvc 응급 키워드 필터 → 감지 시 emergency_fallback 청크만 반환 후 종료.
        (2) 감지 안 되면 User Health Context + RAG Retriever로 컨텍스트 구성 → LLM 스트리밍 호출.
        응답은 `text/plain` 청크 스트림이며, 각 청크는 ChatMessageChunk 구조를 따른다.
        스트림 종료 시 반드시 disclaimer(T-LLM-1)가 포함된 done 청크로 마무리된다.
        비로그인 정책: 이 엔드포인트는 JWT 인증(bearerAuth)이 필수이므로 비로그인 사용자는 애초에 호출할 수 없다
        (프론트는 UI만 노출하고 전송 시 로그인 모달로 유도, 백엔드는 이 인증 요구사항으로 이중 차단).
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: session_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatMessageRequest' }
      responses:
        '200':
          description: 스트리밍 응답 시작 (text/plain, 청크 단위 전송)
          content:
            text/plain:
              schema:
                type: string
                description: "줄바꿈으로 구분된 JSON 청크들 (ChatMessageChunk 참고)"

  # ================= Push 구독 =================
...